+++ scrapers/FL/Bay/Court/scraper/tests/test_ScraperUtils.py
@@ -1,6 +1,6 @@
 import pytest
 from absl.testing import flagsaver
-from common.pii import Pii
+from utils.meta.pii.Pii import Pii
 from scrapers.FL.Bay.Court.scraper.utils import ScraperUtils
 import os
